start splitting polymorphic functions out of the factory
