+++ pkg/kubectl/cmd/clusterinfo_dump.go
@@ -26,14 +26,13 @@ import (
 	"github.com/spf13/cobra"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	restclient "k8s.io/client-go/rest"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
 	api "k8s.io/kubernetes/pkg/apis/core"
 	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
 	"k8s.io/kubernetes/pkg/kubectl/cmd/templates"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions"
+	"k8s.io/kubernetes/pkg/kubectl/polymorphichelpers"
 	"k8s.io/kubernetes/pkg/kubectl/util/i18n"
 	"k8s.io/kubernetes/pkg/printers"
 )
@@ -46,10 +45,11 @@ type ClusterInfoDumpOptions struct {
 	AllNamespaces bool
 	Namespaces    []string
 
-	timeout       time.Duration
-	clientset     internalclientset.Interface
-	namespace     string
-	logsForObject func(object, options runtime.Object, timeout time.Duration) (*restclient.Request, error)
+	Timeout          time.Duration
+	Clientset        internalclientset.Interface
+	Namespace        string
+	RESTClientGetter genericclioptions.RESTClientGetter
+	LogsForObject    polymorphichelpers.LogsForObjectFunc
 
 	genericclioptions.IOStreams
 }
@@ -126,25 +126,27 @@ func (o *ClusterInfoDumpOptions) Complete(f cmdutil.Factory, cmd *cobra.Command)
 	o.PrintFlags.OutputFormat = &jsonOutputFmt
 	o.PrintObj = printer.PrintObj
 
-	o.timeout, err = cmdutil.GetPodRunningTimeoutFlag(cmd)
+	o.Timeout, err = cmdutil.GetPodRunningTimeoutFlag(cmd)
 	if err != nil {
 		return err
 	}
-	o.clientset, err = f.ClientSet()
+	o.Clientset, err = f.ClientSet()
 	if err != nil {
 		return err
 	}
-	o.namespace, _, err = f.DefaultNamespace()
+	o.Namespace, _, err = f.DefaultNamespace()
 	if err != nil {
 		return err
 	}
-	o.logsForObject = f.LogsForObject
+	// TODO this should eventually just be the completed kubeconfigflag struct
+	o.RESTClientGetter = f
+	o.LogsForObject = polymorphichelpers.LogsForObjectFn
 
 	return nil
 }
 
 func (o *ClusterInfoDumpOptions) Run() error {
-	nodes, err := o.clientset.Core().Nodes().List(metav1.ListOptions{})
+	nodes, err := o.Clientset.Core().Nodes().List(metav1.ListOptions{})
 	if err != nil {
 		return err
 	}
@@ -155,7 +157,7 @@ func (o *ClusterInfoDumpOptions) Run() error {
 
 	var namespaces []string
 	if o.AllNamespaces {
-		namespaceList, err := o.clientset.Core().Namespaces().List(metav1.ListOptions{})
+		namespaceList, err := o.Clientset.Core().Namespaces().List(metav1.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -166,14 +168,14 @@ func (o *ClusterInfoDumpOptions) Run() error {
 		if len(o.Namespaces) == 0 {
 			namespaces = []string{
 				metav1.NamespaceSystem,
-				o.namespace,
+				o.Namespace,
 			}
 		}
 	}
 	for _, namespace := range namespaces {
 		// TODO: this is repetitive in the extreme.  Use reflection or
 		// something to make this a for loop.
-		events, err := o.clientset.Core().Events(namespace).List(metav1.ListOptions{})
+		events, err := o.Clientset.Core().Events(namespace).List(metav1.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -181,7 +183,7 @@ func (o *ClusterInfoDumpOptions) Run() error {
 			return err
 		}
 
-		rcs, err := o.clientset.Core().ReplicationControllers(namespace).List(metav1.ListOptions{})
+		rcs, err := o.Clientset.Core().ReplicationControllers(namespace).List(metav1.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -189,7 +191,7 @@ func (o *ClusterInfoDumpOptions) Run() error {
 			return err
 		}
 
-		svcs, err := o.clientset.Core().Services(namespace).List(metav1.ListOptions{})
+		svcs, err := o.Clientset.Core().Services(namespace).List(metav1.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -197,7 +199,7 @@ func (o *ClusterInfoDumpOptions) Run() error {
 			return err
 		}
 
-		sets, err := o.clientset.Extensions().DaemonSets(namespace).List(metav1.ListOptions{})
+		sets, err := o.Clientset.Extensions().DaemonSets(namespace).List(metav1.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -205,7 +207,7 @@ func (o *ClusterInfoDumpOptions) Run() error {
 			return err
 		}
 
-		deps, err := o.clientset.Extensions().Deployments(namespace).List(metav1.ListOptions{})
+		deps, err := o.Clientset.Extensions().Deployments(namespace).List(metav1.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -213,7 +215,7 @@ func (o *ClusterInfoDumpOptions) Run() error {
 			return err
 		}
 
-		rps, err := o.clientset.Extensions().ReplicaSets(namespace).List(metav1.ListOptions{})
+		rps, err := o.Clientset.Extensions().ReplicaSets(namespace).List(metav1.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -221,7 +223,7 @@ func (o *ClusterInfoDumpOptions) Run() error {
 			return err
 		}
 
-		pods, err := o.clientset.Core().Pods(namespace).List(metav1.ListOptions{})
+		pods, err := o.Clientset.Core().Pods(namespace).List(metav1.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -234,7 +236,7 @@ func (o *ClusterInfoDumpOptions) Run() error {
 			writer.Write([]byte(fmt.Sprintf("==== START logs for container %s of pod %s/%s ====\n", container.Name, pod.Namespace, pod.Name)))
 			defer writer.Write([]byte(fmt.Sprintf("==== END logs for container %s of pod %s/%s ====\n", container.Name, pod.Namespace, pod.Name)))
 
-			request, err := o.logsForObject(pod, &api.PodLogOptions{Container: container.Name}, timeout)
+			request, err := o.LogsForObject(o.RESTClientGetter, pod, &api.PodLogOptions{Container: container.Name}, timeout)
 			if err != nil {
 				// Print error and return.
 				writer.Write([]byte(fmt.Sprintf("Create log request error: %s\n", err.Error())))
