+++ pkg/kubectl/cmd/logs_test.go
@@ -18,19 +18,27 @@ package cmd
 
 import (
 	"bytes"
+	"errors"
 	"io/ioutil"
 	"net/http"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/spf13/cobra"
 
+	"fmt"
+
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	restclient "k8s.io/client-go/rest"
 	"k8s.io/client-go/rest/fake"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
 	api "k8s.io/kubernetes/pkg/apis/core"
+	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
 	cmdtesting "k8s.io/kubernetes/pkg/kubectl/cmd/testing"
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions"
+	"k8s.io/kubernetes/pkg/kubectl/polymorphichelpers"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
 )
 
@@ -74,6 +82,16 @@ func TestLog(t *testing.T) {
 			}
 			tf.Namespace = "test"
 			tf.ClientConfigVal = defaultClientConfig()
+			oldLogFn := polymorphichelpers.LogsForObjectFn
+			defer func() {
+				polymorphichelpers.LogsForObjectFn = oldLogFn
+			}()
+			clientset, err := tf.ClientSet()
+			if err != nil {
+				t.Fatal(err)
+			}
+			polymorphichelpers.LogsForObjectFn = logTestMock{client: clientset}.logsForObject
+
 			streams, _, buf, _ := genericclioptions.NewTestIOStreams()
 
 			cmd := NewCmdLogs(tf, streams)
@@ -219,5 +237,22 @@ func TestLogComplete(t *testing.T) {
 		if !strings.Contains(out, test.expected) {
 			t.Errorf("%s: expected to find:\n\t%s\nfound:\n\t%s\n", test.name, test.expected, out)
 		}
+	}
+}
+
+type logTestMock struct {
+	client internalclientset.Interface
+}
+
+func (m logTestMock) logsForObject(restClientGetter genericclioptions.RESTClientGetter, object, options runtime.Object, timeout time.Duration) (*restclient.Request, error) {
+	switch t := object.(type) {
+	case *api.Pod:
+		opts, ok := options.(*api.PodLogOptions)
+		if !ok {
+			return nil, errors.New("provided options object is not a PodLogOptions")
+		}
+		return m.client.Core().Pods(t.Namespace).GetLogs(t.Name, opts), nil
+	default:
+		return nil, fmt.Errorf("cannot get the logs from %T", object)
 	}
 }
