+++ pkg/kubectl/cmd/util/factory.go
@@ -18,7 +18,6 @@ package util
 
 import (
 	"fmt"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
@@ -27,21 +26,17 @@ import (
 
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/watch"
-	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	restclient "k8s.io/client-go/rest"
 	scaleclient "k8s.io/client-go/scale"
 	api "k8s.io/kubernetes/pkg/apis/core"
-	apiv1 "k8s.io/kubernetes/pkg/apis/core/v1"
 	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
-	coreclient "k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset/typed/core/internalversion"
 	"k8s.io/kubernetes/pkg/kubectl"
 	"k8s.io/kubernetes/pkg/kubectl/cmd/util/openapi"
+	"k8s.io/kubernetes/pkg/kubectl/genericclioptions"
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions/resource"
 	"k8s.io/kubernetes/pkg/kubectl/validation"
 	"k8s.io/kubernetes/pkg/printers"
@@ -67,14 +62,7 @@ type Factory interface {
 // Generally provides discovery, negotiation, and no-dep calls.
 // TODO The polymorphic calls probably deserve their own interface.
 type ClientAccessFactory interface {
-	// Returns a client.Config for accessing the Kubernetes server.
-	ToRESTConfig() (*restclient.Config, error)
-	// Returns interfaces for dealing with arbitrary runtime.Objects.
-	ToRESTMapper() (meta.RESTMapper, error)
-	// Returns a discovery client
-	ToDiscoveryClient() (discovery.CachedDiscoveryInterface, error)
-	// Returns kubeconfig loader
-	ToRawKubeConfigLoader() clientcmd.ClientConfig
+	genericclioptions.RESTClientGetter
 
 	// ClientSet gives you back an internal, generated clientset
 	ClientSet() (internalclientset.Interface, error)
@@ -151,8 +139,6 @@ type ObjectMappingFactory interface {
 	// Returns a Describer for displaying the specified RESTMapping type or an error.
 	Describer(mapping *meta.RESTMapping) (printers.Describer, error)
 
-	// LogsForObject returns a request for the logs associated with the provided object
-	LogsForObject(object, options runtime.Object, timeout time.Duration) (*restclient.Request, error)
 	// Returns a HistoryViewer for viewing change history
 	HistoryViewer(mapping *meta.RESTMapping) (kubectl.HistoryViewer, error)
 	// Returns a Rollbacker for changing the rollback version of the specified RESTMapping type or an error
@@ -191,7 +177,7 @@ type factory struct {
 
 // NewFactory creates a factory with the default Kubernetes resources defined
 // Receives a clientGetter capable of providing a discovery client and a REST client configuration.
-func NewFactory(clientGetter RESTClientGetter) Factory {
+func NewFactory(clientGetter genericclioptions.RESTClientGetter) Factory {
 	clientAccessFactory := NewClientAccessFactory(clientGetter)
 	objectMappingFactory := NewObjectMappingFactory(clientAccessFactory)
 	builderFactory := NewBuilderFactory(clientAccessFactory, objectMappingFactory)
@@ -201,51 +187,6 @@ func NewFactory(clientGetter RESTClientGetter) Factory {
 		ObjectMappingFactory: objectMappingFactory,
 		BuilderFactory:       builderFactory,
 	}
-}
-
-// GetFirstPod returns a pod matching the namespace and label selector
-// and the number of all pods that match the label selector.
-func GetFirstPod(client coreclient.PodsGetter, namespace string, selector string, timeout time.Duration, sortBy func([]*v1.Pod) sort.Interface) (*api.Pod, int, error) {
-	options := metav1.ListOptions{LabelSelector: selector}
-
-	podList, err := client.Pods(namespace).List(options)
-	if err != nil {
-		return nil, 0, err
-	}
-	pods := []*v1.Pod{}
-	for i := range podList.Items {
-		pod := podList.Items[i]
-		externalPod := &v1.Pod{}
-		apiv1.Convert_core_Pod_To_v1_Pod(&pod, externalPod, nil)
-		pods = append(pods, externalPod)
-	}
-	if len(pods) > 0 {
-		sort.Sort(sortBy(pods))
-		internalPod := &api.Pod{}
-		apiv1.Convert_v1_Pod_To_core_Pod(pods[0], internalPod, nil)
-		return internalPod, len(podList.Items), nil
-	}
-
-	// Watch until we observe a pod
-	options.ResourceVersion = podList.ResourceVersion
-	w, err := client.Pods(namespace).Watch(options)
-	if err != nil {
-		return nil, 0, err
-	}
-	defer w.Stop()
-
-	condition := func(event watch.Event) (bool, error) {
-		return event.Type == watch.Added || event.Type == watch.Modified, nil
-	}
-	event, err := watch.Until(timeout, w, condition)
-	if err != nil {
-		return nil, 0, err
-	}
-	pod, ok := event.Object.(*api.Pod)
-	if !ok {
-		return nil, 0, fmt.Errorf("%#v is not a pod event", event)
-	}
-	return pod, 1, nil
 }
 
 func makePortsString(ports []api.ServicePort, useNodePort bool) string {
