+++ test/e2e/kubectl/BUILD
@@ -15,7 +15,7 @@ go_library(
     importpath = "k8s.io/kubernetes/test/e2e/kubectl",
     deps = [
         "//pkg/controller:go_default_library",
-        "//pkg/kubectl/cmd/util:go_default_library",
+        "//pkg/kubectl/polymorphichelpers:go_default_library",
         "//test/e2e/framework:go_default_library",
         "//test/e2e/generated:go_default_library",
         "//test/e2e/scheduling:go_default_library",
